Added stuff for special zombie support
Constants for each kind of zombie: chance and wave
Global for wave counter
Chance is taken as "weight": the real chance is the constant divided by the sum of all constants for the zombies for the current wave. set to 10 for testing purposes.
Wave is set as 1 for testing purposes, so specials appear soon in the game. Definitive value should be higher. Note that normal zombies don't need wave, as they start in wave 1.
Wave counter already works, but is not used, as specials spawn is not yet set.